# perf: skip debug work when not in debug mode