`MergedFileDescriptors` was unconditionally checking import paths and file descriptor differences, then conditionally returning the differences as an error. There is no need to check the differences if we will never return them.

This made a measurable difference in running simd q:

```console
$ hyperfine --warmup 4 -r 50 '/tmp/simd-dodebugwork q' '/tmp/simd-skipdebugwork q' Benchmark 1: /tmp/simd-dodebugwork q
  Time (mean ± σ):     121.6 ms ±   6.5 ms    [User: 173.6 ms, System: 23.7 ms]
  Range (min … max):   117.0 ms … 162.7 ms    50 runs

  Warning: Statistical outliers were detected. ...

Benchmark 2: /tmp/simd-skipdebugwork q
  Time (mean ± σ):     113.6 ms ±   6.5 ms    [User: 158.5 ms, System: 23.1 ms]
  Range (min … max):   109.3 ms … 139.7 ms    50 runs

  Warning: Statistical outliers were detected. ...

Summary
  '/tmp/simd-skipdebugwork q' ran
    1.07 ± 0.08 times faster than '/tmp/simd-dodebugwork q'
```

And on two single runs of `TestAppStateDeterminism`, this optimization reduced local runtime from 20.9 seconds to 19.7, an anecdotal ~6% speedup.